Disclaimer about the number of slots.
This is debatable, but this may discourrage students who picked Git more
or less at random. If I read
https://code.google.com/p/google-summer-of-code/wiki/ProgramStatistics
correctly, the average number of slots per organization is closer to 10
than 2, hence students have better chance to get accepted in other
organizations.

OTOH, 2 slots is not so much worse than the 3 we had last year. And now
that Dscho is volunteer to mentor a GSoC too, we may actually end up
accepting 3 students.