revnews: remove meaningless "X also reviewed"
The fact that the patch was reviewed by the maintainer does not add
much value to the section by itself. It is a different matter if it
were accompanied by a summary of value the review added to the
discussion, but otherwise it only makes the section longer without
adding much value ot the reader.