rev_news: redirect rev_news/ to current edition
This makes https://git.github.io/rev_news/ do something
sensible (as opposed to giving a 404). An obvious
alternative would be to just have an index page that links
to the editions in reverse chronological order.

This `redirect_from` will need to be updated manually as new
editions are published.