rev_news: remove legacy url redirect
Remove the redirect-from as it is a bit of an ugly pattern that
generates empty bullet points in the navmenu as it is now.

Instead of redirecting the old url /rev_news/edition-1.html to go to the
first post on /rev_news/2015/03/25/edition-1/ - simply recreate the
edition-1.html page as a copy of the first post. Ugly duplication, but
it's the simplest solution without having webserver-level rewrite
capability.

Ref discussion in #30.