rev_news: remove duplicate posts
The first edition-1.md file was only introduced/kept to maintain the URL
for the original 1st edition, as this was created/named before we got
dynamic date-based edition URLs. There is no need to duplicate the 2nd,
3rd, or Nth edition posting.

See also commit 31c1a03978f8 which introduced the edition-1.md
duplicate.

Closes #64.