rev_news: remove some meta-data from posts
Since we only have one blog post category, there's no sense in showing
this at the top of each edition. Also, we don't like the way author is
displayed near top, as we prefer noting the curators at the bottom of
each edition.

This takes care of Christian's complaint in #45.