add google analytics
See issue #57

We want to be able to track the popularity of the Git Rev News. Hence,
adding some sort of tracking would be valuable and motivational. Perhaps
some people don't like the notion of Google tracking traffic, but it is
currently the de-facto way of measuring internet traffic, and it
requires no server-side modification, which would otherwise be difficult
on GitHub Pages/Jekyll.

The initial tracking is made to my own Google Analytics account, but
once it works I'll try to find out how it can be shared with other Git
Rev News editors.