2017-microprojects: convert $gmane references to links
We do not define the `$gmane` construct here at all. It's
well known to list regulars, but probably not to applicants.
Moreover, gmane is now defunct, so you have to know to dig
up the link via public-inbox.

Let's convert these to public-inbox references. And since
this is real HTML, we can even make them hyperlinks.
Amazing!