navbar: skip pages with no title
We seem to get a bogus empty bullet in the navbar due to
"style.css" being listed. I have a feeling there's a more
elegant fix here (like filtering out things that aren't
"real" pages from the site.pages list). But as a workaround,
let's just skip anything with no title. That's sane to do
anyway, even for real pages.

Oddly, this bug doesn't seem to replicate on my local
"jekyll serve", but it is present on the published site.