navbar: reject entries with nil titles rather than empty
The blank entry in the navbar was supposed to be fixed by
c7afa1111e64c0b65c4e481c00ea21ac6a405eff, but it didn't
work. This is presumably because the title there is nil, not
the empty string (but I can't reproduce locally, so I'm just
guessing).

Let's switch out the empty-string test for a "truthy" one
(technically we could do both, but if the only way to get an
empty string is by explicitly saying "title: " in the front
matter, it doesn't seem likely to happen).