rn-27: Reorganize and extend 'cregit'-related links
The LWN.net article link was already posted in edition 19, but
at that time 'cregit' tool was not open-sourced, and had yet to
get this name.  Reorganize and extend information about this tool.

Note: the slides needs to be linked, as the tool is underdocumented
in its GitHub repository.

If you have better idea how to present those links (for example
avoid sub-lists), feel free to change it.