add landing page for outreachy round 15
We're participating in Outreachy this round, so let's clean
up the old landing page and give some project ideas.

I've dropped some of the details from our landing page in
favor of just linking to the Outreachy site, which is pretty
well organized.

For the projects, I ported over items from the SoC ideas
page that I thought could easily be mentored by me or
Christian.