remove stash'ing merge conflicts project
This project depends on stash-in-c going to at least `next`.  I'm
going to send an updated version of the patch series tonight, but
I'm not very confident at this point that it's actually going to
make it to next before GSoC starts.  I'd rather take the project
away now, before applications actually open, and later
re-introduce it, than misleading a student.