SoC-2019-Ideas: don't split text within the `s
When the text within the ` (tildes)  is split between two lines
it is not rendered as expected. To be precise the part that
comes in the second line isn't visible.

It's worth noting that this is not clear from the rendering
of the markdown by GitHub. But it's clear from the syntax
highlighting done by the editor in GitHub and the rendering
in the website (where this was first noted).