SoC-2019-Ideas: add two more ideas
Those are smaller projects, so that Git can accept students from a wider
audience (i.e. students who are not already familiar with C, or with
mailing list-centered project workflows). The downside is that this will
require more effort from the mentors than was common in past years. The
upside is that this will bring all the benefits of more inclusive
development.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>