SoC-2020-Ideas: Add 'Generation number v2' idea
Derrick Stolee recommended that a GSoC student update the generation
number to "v2" instead of working on min-post intervals labeling that
might not get merged in (and is more of a research project, than
engineering one).

Christian Couder asked for this idea to be merged into "Graph labeling
for speeding up git commands" project idea instead of adding new
project ideas at that time.

That is why 'Generation number v2' got added as first of alternative
tasks to solve for the 'Graph labeling' general idea.