Hacking-Git: use a known version for link
In c271b5d (Hacking-Git: highlight correct lines in CodingGuidelines,
2020-06-07), we fixed lines which are highlighted in the link to the
"Launching gdb explanations in CodingGuidelines" section. However, this
is liable to break in the future as the document is updated.

Hardcode that link to point to the v2.27.0 release of Git instead of the
latest 'master' so that the selected lines will never change.

Although this may potentially present stale information to readers, this
is mitigated by two things. First, I don't anticipate that the GDB
instructions are likely to change. Second, in the section above, there
is a link to the 'master' version of CodingGuidelines so users looking
for that will find a link to the "fresh" version.