css: move the max-width to the site
This allows the wrapper to fit the whole width of the window, and align
the contained elements properly.

The width of the navbar is 20em (280px), so the max-width of the site is
940px - 280px, thus retaining the original width.

This fixes the issue with "Git Rev News: Edition 68" where the site width
was too big, and the navbar width got too small.

Signed-off-by: Felipe Contreras <felipe.contreras@gmail.com>