Mentoring-Guide: use proper syntax for ordered list
The syntax used for an ordered list isn't one that is recognized
by the Markdown parser used for the generated site. So, the
list is rendered in HTML as though it is not a list thus needing
change done in bda6f86 (Mentoring-Program-Guide: make lines break,
2021-06-14).

Change the syntax for the list to use the proper Markdown syntax
for an ordered list.

As this makes the fix mentioned in the commit above unnecessary,
that change has been reverted too.