SoC-2022-Ideas.md: add "reachability bitmap improvements"
As discussed[1] on the list, here is a description of a GSoC project
which takes a look at a handful of potential performance improvements
for reachability bitmaps.

Like my Outreachy proposal[2], there isn't one specific bitmap project,
but instead a handful of small- to medium-sized sub-projects that
students should feel free to choose from. This is by design to protect
against the case that one project doesn't carry the benefits we expect,
or a student changes their interest, etc.

[1]: https://lore.kernel.org/git/YfGpjm8mCkWIPM6V@nand.local/
[2]: https://www.outreachy.org/outreachy-december-2021-internship-round/communities/git/#improve-reachability-bitmaps