Include development info and re-shuffle home/about pages
The information to spin up the site locally got buried inside
_includes/README in a recent change. It actually belongs to the
README present in the root of the repo, too since that is what users
see when viewing the repository. Hence, move that information there.

The "About" page has long been very identical to the home page. So,
shuffle the content in home and about pages and unify them into
the Home page itself. For historical purposes, make the about page
redirect to the homepage.

Fixes #610