SoC-2024: improve positioning of logo in small screens
We could align the image to right only on bigger screens as it looks
good only when the space on the left is enough for some decent amount
of content to be shown. On smaller screens, we could just place the
image on its one chunk of space (default behaviour when including
images in Markdown) and let the content follow it.

Tweak the styles to accordingly for this.