GSoC-participants: add file with  partial info about participants
This file is being collected to gather the information about the
participants that have contributed to Git via GSoC, an outreach
program conducted by Google each year. This would serve as a helpful
reference.

The intention is not to restrict gathering the information only for
GSoC. We've started gather this with GSoC. We'll expand to also gather
information of participants who've contributed via Outreachy.