rn-116: minor copy edit
I'm also overriding some previous edits, so perhaps some
explanation is in order:

1. 079a323e4a75 changed some 'open-source' to 'open source',
but not all and not consistently.  Standard English uses
dashed forms for adjectives and non-dashed forms for nouns;
follow that rule.

2. 079a323e4a75 changed "people's" to "peoples'", which is
not correct here (it would imply a number of ethnic groups
is meant, but that's not the case); restore the original.

(As an aside, personally I tend to leave the interviews
alone beyond obvious typos, as I consider the language
idiosyncrasies part of the glimpse into an interviewee's
personality.)